+++ tests/gtests/sw_testhelpers.cc
@@ -24,8 +24,17 @@
 #include "include/SW_Files.h"
 #include "include/myMemory.h"
 
+#include "include/SW_Weather.h"
+#include "include/SW_Main_lib.h"
+
+
 #include "tests/gtests/sw_testhelpers.h"
 
+
+void silent_tests(LOG_INFO* LogInfo) {
+  LogInfo->logged = swFALSE;
+  LogInfo->logfp = NULL;
+}
 
 /**
   @brief Creates soil layers based on function arguments (instead of reading
@@ -38,7 +47,7 @@
       simulation after this function has set soil layers, e.g.,
       SW_SWC_init_run()
 */
-void create_test_soillayers(unsigned int nlayers, char *InFiles[],
+void create_test_soillayers(unsigned int nlayers,
       SW_VEGPROD *SW_VegProd, SW_SITE *SW_Site, LOG_INFO *LogInfo) {
 
   if (nlayers <= 0 || nlayers > MAX_LAYERS) {
@@ -95,5 +104,70 @@ void create_test_soillayers(unsigned int nlayers, char *InFiles[],
   set_soillayers(SW_VegProd, SW_Site, nlayers, dmax,
     bulkd, f_gravel, evco, trco_grass, trco_shrub, trco_tree,
     trco_forb, psand, pclay, imperm, soiltemp, nRegions,
-    regionLowerBounds, InFiles, LogInfo);
+    regionLowerBounds, LogInfo);
+}
+
+
+
+void setup_SW_Site_for_tests(SW_SITE *SW_Site) {
+    LOG_INFO LogInfo;
+    silent_tests(&LogInfo);
+
+    SW_Site->deepdrain = swTRUE;
+
+    SW_Site->_SWCMinVal = 100;
+    SW_Site->_SWCWetVal = 15;
+    SW_Site->_SWCInitVal = 15;
+
+    SW_Site->stMaxDepth = 990;
+    SW_Site->stDeltaX = 15;
+
+    SW_Site->slow_drain_coeff = 0.02;
+
+    SW_Site->site_has_swrcp = swFALSE;
+    strcpy(SW_Site->site_swrc_name, (char *) "Campbell1974");
+    SW_Site->site_swrc_type = encode_str2swrc(SW_Site->site_swrc_name, &LogInfo);
+    strcpy(SW_Site->site_ptf_name, (char *) "Cosby1984AndOthers");
+    SW_Site->site_ptf_type = encode_str2ptf(SW_Site->site_ptf_name);
+}
+
+
+// `memcpy()` does not work for copying an initialized `SW_ALL`
+// because it does not copy dynamically allocated memory to which
+// members of `SW_ALL` point to
+void setup_AllTest_for_tests(
+  SW_ALL *SW_All,
+  PATH_INFO *PathInfo,
+  LOG_INFO *LogInfo,
+  SW_OUTPUT_POINTERS *SW_OutputPtrs
+) {
+      Bool QuietMode = swFALSE;
+
+      // Initialize SOILWAT2 variables and read values from example input file
+      silent_tests(LogInfo);
+
+      PathInfo->InFiles[eFirst] = Str_Dup(DFLT_FIRSTFILE, LogInfo);
+
+      SW_CTL_setup_model(SW_All, SW_OutputPtrs, PathInfo, LogInfo);
+      SW_CTL_read_inputs_from_disk(SW_All, PathInfo, LogInfo);
+
+      /* Notes on messages during tests
+        - `SW_F_read()`, via SW_CTL_read_inputs_from_disk(), writes the file
+          "example/Output/logfile.log" to disk (based on content of "files.in")
+        - we close "Output/logfile.log"
+        - we set `logfp` to NULL to silence all non-error messages during tests
+        - error messages go directly to stderr (which DeathTests use to match against)
+      */
+      sw_check_log(QuietMode, LogInfo);
+      silent_tests(LogInfo);
+
+      SW_WTH_finalize_all_weather(
+        &SW_All->Markov,
+        &SW_All->Weather,
+        SW_All->Model.cum_monthdays,
+        SW_All->Model.days_in_month,
+        LogInfo
+      );
+
+      SW_CTL_init_run(SW_All, LogInfo);
 }