+++ tests/gtests/test_SW_Defines.cc
@@ -23,7 +23,7 @@
 
 namespace {
 
-  TEST_F(AllTest, SWDefinesMissingValues) {
+  TEST(SWDefines, SWDefinesMissingValues) {
     // SOILWAT2 missing value
     EXPECT_TRUE(missing(SW_MISSING));